+++ Coditech.Project/API/Coditech.Engine.HospitalManagementSystem/Service/Interface/IHospitalDoctorOPDScheduleService.cs
@@ -8,7 +8,7 @@ namespace Coditech.API.Service
     public interface IHospitalDoctorOPDScheduleService
     {
         HospitalDoctorOPDScheduleListModel GetHospitalDoctorOPDScheduleList(string selectedCentreCode, short selectedDepartmentId, FilterCollection filters, NameValueCollection sorts, NameValueCollection expands, int pagingStart, int pagingLength);
-        HospitalDoctorOPDScheduleModel GetHospitalDoctorOPDSchedule(int hospitalDoctorId,int hospitalDoctorOPDScheduleId);
+        HospitalDoctorOPDScheduleModel GetHospitalDoctorOPDSchedule(int hospitalDoctorId,long hospitalDoctorOPDScheduleId);
         bool UpdateHospitalDoctorOPDSchedule(HospitalDoctorOPDScheduleModel model);
     }
 }
